# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenWrapper Gateway API",
    "version": "0.1.4",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
...
  },
  "servers": [
    {
      "url": "https://gateway.openwrapper.muejam.com",
      "description": "Production OpenWrapper Gateway (TLS - Paymob, Fawry, Stripe)"
    },
    {
      "url": "https://openwrapper.muejam.com/api",
      "description": "Production Web API (Control Plane, Better Auth, & Ledger GraphQL)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Rust gateway (Paymob, Fawry, Stripe, gRPC & GraphQL)"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Web Control Plane & API"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/payments/{id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Get payment status",
        "operationId": "getPayment",
        "description": "Returns the current payment record. For an `unknown` payment,\nthe gateway may perform one provider status inquiry before responding.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/PaymentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Current payment record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID or malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Payment not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "OpenWrapper API key in `Authorization: Bearer <key>`."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "OpenWrapper API key sent directly in `X-API-Key`."
      }
    },
    "parameters": {
...
      "PaymentId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Opaque OpenWrapper payment identifier.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "example": "01J6G7RPQZ4Z9M6W3V2T1K8N90"
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing/invalid API key or rejected provider credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Caller is not authorized for this operation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Gateway or provider rate limit exceeded.",
        "headers": {
          "Retry-After": {
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal gateway failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadGateway": {
        "description": "Provider or downstream gateway failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "GatewayTimeout": {
        "description": "Provider operation exceeded the gateway deadline.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "PaymentProvider": {
        "type": "string",
        "enum": [
          "paymob",
          "fawry",
          "stripe",
          "mock"
        ],
        "description": "Provider adapter. Native first-class support for `paymob`, `fawry`,\n`stripe`, and `mock` across both the Rust gateway and web API.\n"
      },
      "PaymentCurrency": {
        "type": "string",
        "enum": [
          "EGP",
          "USD",
          "EUR",
          "GBP",
          "SAR",
          "AED",
          "KWD",
          "BHD",
          "OMR",
          "JPY"
        ],
        "description": "ISO-4217 currency code. Paymob and Fawry operate primarily in EGP;\nStripe and Mock support multi-currency payments.\n"
      },
...
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "succeeded",
          "failed",
          "unknown"
        ],
        "description": "`unknown` is a recorded ambiguous provider outcome, not an API error.\nIt must be reconciled by status inquiry or an authenticated webhook.\n"
      },
      "PaymentNextAction": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/RedirectToUrlAction"
          },
          {
            "$ref": "#/components/schemas/PayAtReferenceAction"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "redirect_to_url": "#/components/schemas/RedirectToUrlAction",
            "pay_at_reference": "#/components/schemas/PayAtReferenceAction"
          }
        }
      },
      "RedirectToUrlAction": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "url"
        ],
        "properties": {
          "type": {
            "type": "string",
            "const": "redirect_to_url"
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "PayAtReferenceAction": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "reference"
        ],
        "properties": {
          "type": {
            "type": "string",
            "const": "pay_at_reference"
          },
          "reference": {
            "type": "string"
          },
          "instructions": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "PaymentView": {
        "type": "object",
        "required": [
          "payment_id",
          "provider",
          "provider_reference",
          "status",
          "amount_minor_units",
          "currency",
          "merchant_reference"
        ],
        "properties": {
          "payment_id": {
            "type": "string",
            "minLength": 1
          },
          "provider": {
            "$ref": "#/components/schemas/PaymentProvider"
          },
          "provider_reference": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "amount_minor_units": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 1000000000
          },
          "currency": {
            "$ref": "#/components/schemas/PaymentCurrency"
          },
          "merchant_reference": {
            "type": [
              "string",
              "null"
            ]
          },
          "next_action": {
            "$ref": "#/components/schemas/PaymentNextAction"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error category.",
            "enum": [
              "invalid_request",
              "missing_provider_credentials",
              "idempotency_conflict",
              "unauthorized",
              "not_found",
              "unsupported_provider",
              "gateway_required",
              "gateway_unavailable",
              "gateway_unreachable",
              "gateway_error",
              "validation_error",
              "authentication_error",
              "authorization_error",
              "configuration_error",
              "network_error",
              "timeout",
              "provider_error",
              "rate_limit",
              "unsupported_capability",
              "security_error",
              "unknown_outcome",
              "internal_error"
            ]
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": true,
            "description": "Optional validation or error details."
          }
        }
      },
...
    }
  }
}